move train to scripts. make sample rate a float

+++ umx/data.py
@@ -137,7 +137,7 @@ class UnmixDataset(torch.utils.data.Dataset):
     def __init__(
             self,
             root: Union[Path, str],
-            sample_rate: int,
+            sample_rate: float,
             seq_duration: Optional[float] = None,
             source_augmentations: Optional[Callable] = None,
     ) -> None:
@@ -366,7 +366,7 @@ class AlignedDataset(UnmixDataset):
         output_file: str = 'vocals.wav',
         seq_duration: Optional[float] = None,
         random_chunks: bool = False,
-        sample_rate: int = 44100,
+        sample_rate: float = 44100.0,
         source_augmentations: Optional[Callable] = None,
         seed: int = 42
     ) -> None:
@@ -455,7 +455,7 @@ class SourceFolderDataset(UnmixDataset):
         nb_samples: int = 1000,
         seq_duration: Optional[float] = None,
         random_chunks: bool = True,
-        sample_rate: int = 44100,
+        sample_rate: float = 44100.0,
         source_augmentations: Optional[Callable] = lambda audio: audio,
         seed: int = 42
     ) -> None:
@@ -557,7 +557,7 @@ class FixedSourcesTrackFolderDataset(UnmixDataset):
         random_chunks: bool = False,
         random_track_mix: bool = False,
         source_augmentations: Optional[Callable] = lambda audio: audio,
-        sample_rate: int = 44100,
+        sample_rate: float = 44100.0,
         seed: int = 42
     ) -> None:
         """A dataset that assumes audio sources to be stored
@@ -687,7 +687,7 @@ class VariableSourcesTrackFolderDataset(UnmixDataset):
         seq_duration: Optional[float] = None,
         random_chunks: bool = False,
         random_interferer_mix: bool = False,
-        sample_rate: int = 44100,
+        sample_rate: float = 44100.0,
         source_augmentations: Optional[Callable] = lambda audio: audio,
         silence_missing_targets: bool = False
     ) -> None:
@@ -892,7 +892,7 @@ class MUSDBDataset(UnmixDataset):
             download=download,
             *args, **kwargs
         )
-        self.sample_rate = 44100  # musdb is fixed sample rate
+        self.sample_rate = 44100.0  # musdb is fixed sample rate
 
     def __getitem__(self, index):
         audio_sources = []
